wip moving the reads io object out
